patient-details-view, favicon, layout-cleanup, breadcrumb, file-rename, styles
