+++ application/helper.py
@@ -1,0 +1,6 @@
+import uuid
+
+def get_unique_id():
+  return uuid.uuid1().hex
+
+print(get_unique_id())
