+++ application/routes/auth.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
-from application.models import User
+from application.model import User
 
 blueprint = Blueprint('auth_bp', __name__, url_prefix='/auth')
