+++ application/static/css/custom.css
@@ -1,0 +1,5 @@
+.details-view-avatar {
+  max-width: 175px;
+  margin-top: 20px;
+  margin-left: 20px;
+}
