+++ application/templates/home/index.html
@@ -1,17 +1,27 @@
 {% extends 'layout/layout.html' %}
+{% from 'layout/helper.html' import breadcrumb %}
 
 {% block style %} {% endblock style %}
+
 {% block javascript %} {% endblock javascript %}
 
-{% block title %}Boilerplate{% endblock title %}
+{% block title %} Boilerplate {% endblock title %}
 
-{% block breadcrumb %}Boilerplate{% endblock breadcrumb %}
+
+{% block breadcrumb %}
+{{ breadcrumb({'Boiler':  '/', 'Plate': ''} ) }}
+{% endblock breadcrumb %}
+
+
 
 {% block content %}
-<div class="h1 mb-4">Boilerplate</div>        
+<div class="h2 mb-4 mt-1">Boilerplate</div>
+
 <div class="row">
   <div class="col">
-    Boilerplate
+    {{ lipsum(n=5, html=True, min=20, max=100) }}
   </div>
 </div>
+
+
 {% endblock content %}
