+++ application/templates/layout/layout.html
@@ -7,9 +7,21 @@
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
   <meta name="description" content="Medico" />
   <meta name="author" content="Asif R. Porosh" />
-  <title>{% block title %}{% endblock title%}</title>
-  <link href="../static/css/styles.css" rel="stylesheet" />
+
+  <link rel="apple-touch-icon" sizes="180x180" href='{{ url_for("static", filename="img/favicon/apple-touch-icon.png" ) }}'>   
+  <link rel="icon" type="image/png" sizes="32x32" href='{{ url_for("static", filename="img/favicon/favicon-32x32.png" ) }}'>  
+  <link rel="icon" type="image/png" sizes="16x16" href='{{ url_for("static", filename="img/favicon/favicon-16x16.png" ) }}'>  
+  <link rel="manifest" href='{{ url_for("static", filename="img/favicon/site.webmanifest" ) }}'>  
+  <link rel="mask-icon" href='{{ url_for("static", filename="img/favicon/safari-pinned-tab.svg" ) }}' color='#5bbad5'>   
+  <link rel="shortcut icon" href='{{ url_for("static", filename="img/favicon/favicon.ico" ) }}'>
+  <meta name="msapplication-TileColor" content="#ffc40d">
+  <meta name="msapplication-config" content="{{ url_for("static", filename="img/favicon/browserconfig.xml" ) }}">
+  <meta name="theme-color" content="#ffffff">
+
+  <link href='{{ url_for("static", filename="css/styles.css") }}' rel="stylesheet" />
+  <link href='{{ url_for("static", filename="css/custom.css") }}' rel="stylesheet" />
   
+  <title>{{config.APP_NAME}} - {% block title %}{% endblock title%}</title>
   {% block style %}{% endblock style %}
 </head>
 
@@ -22,9 +34,7 @@
     <div id="layoutSidenav_content">
       <main>
         <div class="container-fluid px-4">
-          <ol class="breadcrumb mt-2 mb-2">
-            <li class="breadcrumb-item active">{% block breadcrumb %}{% endblock breadcrumb %}</li>
-          </ol>
+          {% block breadcrumb %}{% endblock breadcrumb %}
           {% block content %}{% endblock content%}
         </div>
       </main>
@@ -33,7 +43,7 @@
   </div>
   <script src="https://use.fontawesome.com/releases/v6.1.0/js/all.js" crossorigin="anonymous"></script>
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
-  <script src="../static/js/scripts.js"></script>
+  <script src='{{ url_for("static", filename="js/scripts.js")}}'></script>
 
   {% block javascript %}{% endblock javascript %}
 </body>
